Allow libovsdb to connnect to ovsdb instance over IPv6 connection.
When trying to connect to an ovsdb instance via IPv6, encountered an IPv6
parsing issue. In the logic to strip off the leading protocol (i.e. tcp:...),
the code uses strings.Split(endpoint, ":"), which does not take into account
the ':' in an IPv6 address.

Updated the code to use the existing URL parsed data.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>