schema: Add full support for OVSDB types
The way the column types are specified in RFC7047 is not directly
unmarshallable by the json encoder.

This patch adds the necessary structs and functions to manually
unmarshal the schema definition.

The result is the possibility of determining the exact time of golang
structure needed to store each column type.

Several approaches were considering when implementing this:
A) Multiple rounds of json.Unmarshal storing the partial in in-struct
json.rawMessage
B) Single call to json.Unmarshall and manually decoding the rest of the
message (using reflex to guess the type of incoming data)
C) Multiple rounds of json.Unmarshal storing partials in private
temporary structs.

After discarting A) for it's high memory consumption, B) and C) were
compared:
B) has lower memory allocation needs but more LOC
C) has higher memory allocation needs (though after GC, it's the same or
less than B), and fewer LOC.

This patch implements C) as a good balance between maintainablity and
memory efficiency.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>