client: Support a single database per client
While RFC 7047 allows for a database server to have multiple databases,
in the wild, this is not the case for Open vSwitch or OVN.

There are cases where supporting multiple databases would be odd.
For example, if you were to Monitor the address_set table of the OVN
North and South databases, the client would get no indication in the
corresponding update notifiction of which DB the update was for.

As such, it seems sensible to only support one database per client

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>